Cover neutral add-in page composition
